Moved libspotify session object to its own class
Starting point for a module that would handle the communication to
libspotify. The idea is to eventually make a barrier behind this
module - no libspotify internals should be revealed to the users
of this module.

In this commit the sp_session object creation is put into this module.
The created object can still be fetched with the Session::session()
method since it's used in many places. But that method should be
removed eventually.