Moved the session callbacks to the wrapper
The user now gets Qt signals from the module for the various events
happening in the session. The audio playing is handled via a direct
callback since it needs to return a value so a queued signal connection
wouldn't work.

Introduced an internal SessionMaster class into the wrapper module.
There is one instance of this class per spotify session (although there
can be only one session in libspotify currently). This class makes sure
that the user's correct session listeners get the notifications when
something happens in the session.

The SessionMaster class is also responsible for telling libspotify to
process its events when it needs to do so. This task is now completely
hidden from the user and happens automatically without the user's
action of any kind.